+++ client/www/app/getadb/[...slug]/page.tsx
@@ -1,0 +1,5 @@
+import { notFound } from 'next/navigation';
+
+export default function GetadbCatchAll() {
+  notFound();
+}
