+++ client/www/app/getadb/human/page.tsx
@@ -1,6 +1,9 @@
 export const metadata = {
   title: 'getadb — for humans',
   description: 'Provision an Instant app from your agent.',
+  alternates: {
+    canonical: 'https://getadb.com/human',
+  },
 };
 
 export default function GetadbHumanPage() {
@@ -15,6 +18,13 @@ export default function GetadbHumanPage() {
       <pre className="mt-2 rounded bg-neutral-100 p-4 text-sm overflow-x-auto">
         curl -L &apos;https://getadb.com/?title=my-app&apos;
       </pre>
+      <p className="mt-6 text-sm text-neutral-600">
+        If you&apos;re an agent, load{' '}
+        <a className="underline" href="https://getadb.com/AGENTS.md">
+          https://getadb.com/AGENTS.md
+        </a>{' '}
+        instead.
+      </p>
     </main>
   );
 }
